requirements.txt
pandas>=2.0
pyarrow>=14.0
requests>=2.31
google-cloud-bigquery>=3.13
google-generativeai>=0.5
sentence-transformers>=2.7
hdbscan>=0.8.33
scipy>=1.12
numpy>=1.26
pyyaml>=6.0
fredapi>=0.5
matplotlib>=3.8
pytest>=8.0
